closes #1177
