+++ pymc_marketing/mmm/linear_trend.py
@@ -67,7 +67,7 @@ from pymc.distributions.shape_utils import Dims
 from pytensor.tensor.variable import TensorVariable
 from typing_extensions import Self
 
-from pymc_marketing.mmm.plot import SelToString, plot_curve
+from pymc_marketing.plot import SelToString, plot_curve
 from pymc_marketing.prior import Prior, create_dim_handler
